Generate command test files
A script that can help to generate the command tests files so we do not need to craft them manually.
The only manual step will be the process of reviewing them to make sure the output is reasonable.

Test Plan:
- Check out the branch, run `./Scripts/generate_tests_commands_files.py`, make sure the script does not fail

+++ Sources/XCDiff/main.swift
@@ -21,7 +21,7 @@ func main() -> Int32 {
     let arguments = Array(ProcessInfo.processInfo.arguments.dropFirst())
     let commandRunner = CommandRunner()
     let code = commandRunner.run(with: arguments)
-    return Int32(code)
+    return code
 }
 
 exit(main())
